REF: Move some enum functions to separate file
