+++ rust/dbn/src/enums.rs
@@ -1,6 +1,7 @@
+//! Enums used in Databento APIs.
 #![allow(deprecated)] // TODO: remove with SType::Smart
 
-//! Enums used in Databento APIs.
+mod methods;
 
 use std::fmt::{self, Display, Formatter};
 
@@ -136,31 +137,6 @@ pub enum InstrumentClass {
 impl From<InstrumentClass> for char {
     fn from(class: InstrumentClass) -> Self {
         u8::from(class) as char
-    }
-}
-
-impl InstrumentClass {
-    /// Returns `true` if the instrument class is a type of option.
-    ///
-    /// Note: excludes [`Self::MixedSpread`], which *may* include options.
-    pub fn is_option(&self) -> bool {
-        matches!(self, Self::Call | Self::Put | Self::OptionSpread)
-    }
-
-    /// Returns `true` if the instrument class is a type of future.
-    ///
-    /// Note: excludes [`Self::MixedSpread`], which *may* include futures.
-    pub fn is_future(&self) -> bool {
-        matches!(self, Self::Future | Self::FutureSpread)
-    }
-
-    /// Returns `true` if the instrument class is a type of spread, i.e. composed of two
-    /// or more instrument legs.
-    pub fn is_spread(&self) -> bool {
-        matches!(
-            self,
-            Self::FutureSpread | Self::OptionSpread | Self::MixedSpread
-        )
     }
 }
 
@@ -370,8 +346,6 @@ pub mod rtype {
     use dbn_macros::MockPyo3;
     use num_enum::TryFromPrimitive;
 
-    use super::Schema;
-
     /// A [record type](https://databento.com/docs/standards-and-conventions/common-fields-enums-types),
     /// i.e. a sentinel for different types implementing [`HasRType`](crate::record::HasRType).
     ///
@@ -512,62 +486,6 @@ pub mod rtype {
     pub const BBO_1S: u8 = RType::Bbo1S as u8;
     /// Denotes a best bid and offer record subsampled on a one-minute interval.
     pub const BBO_1M: u8 = RType::Bbo1M as u8;
-
-    /// Get the corresponding `rtype` for the given `schema`.
-    impl From<Schema> for RType {
-        fn from(schema: Schema) -> Self {
-            match schema {
-                Schema::Mbo => RType::Mbo,
-                Schema::Mbp1 | Schema::Tbbo => RType::Mbp1,
-                Schema::Mbp10 => RType::Mbp10,
-                Schema::Trades => RType::Mbp0,
-                Schema::Ohlcv1S => RType::Ohlcv1S,
-                Schema::Ohlcv1M => RType::Ohlcv1M,
-                Schema::Ohlcv1H => RType::Ohlcv1H,
-                Schema::Ohlcv1D => RType::Ohlcv1D,
-                Schema::OhlcvEod => RType::OhlcvEod,
-                Schema::Definition => RType::InstrumentDef,
-                Schema::Statistics => RType::Statistics,
-                Schema::Status => RType::Status,
-                Schema::Imbalance => RType::Imbalance,
-                Schema::Cmbp1 => RType::Cmbp1,
-                Schema::Cbbo1S => RType::Cbbo1S,
-                Schema::Cbbo1M => RType::Cbbo1M,
-                Schema::Tcbbo => RType::Tcbbo,
-                Schema::Bbo1S => RType::Bbo1S,
-                Schema::Bbo1M => RType::Bbo1M,
-            }
-        }
-    }
-
-    /// Tries to convert the given rtype to a [`Schema`].
-    ///
-    /// Returns `None` if there's no corresponding `Schema` for the given rtype or
-    /// in the case of [`OHLCV_DEPRECATED`], it doesn't map to a single `Schema`.
-    pub fn try_into_schema(rtype: u8) -> Option<Schema> {
-        match rtype {
-            MBP_0 => Some(Schema::Trades),
-            MBP_1 => Some(Schema::Mbp1),
-            MBP_10 => Some(Schema::Mbp10),
-            OHLCV_1S => Some(Schema::Ohlcv1S),
-            OHLCV_1M => Some(Schema::Ohlcv1M),
-            OHLCV_1H => Some(Schema::Ohlcv1H),
-            OHLCV_1D => Some(Schema::Ohlcv1D),
-            OHLCV_EOD => Some(Schema::OhlcvEod),
-            STATUS => Some(Schema::Status),
-            INSTRUMENT_DEF => Some(Schema::Definition),
-            IMBALANCE => Some(Schema::Imbalance),
-            STATISTICS => Some(Schema::Statistics),
-            MBO => Some(Schema::Mbo),
-            CMBP_1 => Some(Schema::Cmbp1),
-            CBBO_1S => Some(Schema::Cbbo1S),
-            CBBO_1M => Some(Schema::Cbbo1M),
-            TCBBO => Some(Schema::Tcbbo),
-            BBO_1S => Some(Schema::Bbo1S),
-            BBO_1M => Some(Schema::Bbo1M),
-            _ => None,
-        }
-    }
 
     impl std::str::FromStr for RType {
         type Err = crate::Error;
@@ -1241,26 +1159,6 @@ pub enum TriState {
     Yes = b'Y',
 }
 
-impl From<TriState> for Option<bool> {
-    fn from(value: TriState) -> Self {
-        match value {
-            TriState::NotAvailable => None,
-            TriState::No => Some(false),
-            TriState::Yes => Some(true),
-        }
-    }
-}
-
-impl From<Option<bool>> for TriState {
-    fn from(value: Option<bool>) -> Self {
-        match value {
-            Some(true) => Self::Yes,
-            Some(false) => Self::No,
-            None => Self::NotAvailable,
-        }
-    }
-}
-
 /// How to handle decoding DBN data from other versions.
 #[derive(Debug, Clone, Copy, Default, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[cfg_attr(
@@ -1281,40 +1179,6 @@ pub enum VersionUpgradePolicy {
     /// fail.
     #[default]
     UpgradeToV3,
-}
-
-impl VersionUpgradePolicy {
-    /// Validates a given DBN `version` is compatible with the upgrade policy.
-    ///
-    /// # Errors
-    /// This function returns an error if the version and upgrade policy are
-    /// incompatible.
-    pub fn validate_compatibility(self, version: u8) -> crate::Result<()> {
-        if version > 2 && self == VersionUpgradePolicy::UpgradeToV2 {
-            Err(crate::Error::decode("Invalid combination of `VersionUpgradePolicy::UpgradeToV2` and input version 3. Choose either `AsIs` and `UpgradeToV3` as an upgrade policy"))
-        } else {
-            Ok(())
-        }
-    }
-
-    pub(crate) fn is_upgrade_situation(self, version: u8) -> bool {
-        match (self, version) {
-            (VersionUpgradePolicy::AsIs, _) => false,
-            (VersionUpgradePolicy::UpgradeToV2, v) if v < 2 => true,
-            (VersionUpgradePolicy::UpgradeToV2, _) => false,
-            (VersionUpgradePolicy::UpgradeToV3, v) if v < 3 => true,
-            (VersionUpgradePolicy::UpgradeToV3, _) => false,
-        }
-    }
-
-    /// Returns the output DBN version given the input version and upgrade policy.
-    pub fn output_version(self, input_version: u8) -> u8 {
-        match self {
-            VersionUpgradePolicy::AsIs => input_version,
-            VersionUpgradePolicy::UpgradeToV2 => 2,
-            VersionUpgradePolicy::UpgradeToV3 => 3,
-        }
-    }
 }
 
 /// An error code from the live subscription gateway.
